Delete several unused or obsolescent modules
