--- src/boututils/idl_tabulate.py
@@ -1,16 +1,0 @@
-import numpy as np
-import scipy
-
-
-def idl_tabulate(x, f, p=5):
-    def newton_cotes(x, f):
-        if x.shape[0] < 2:
-            return 0
-        rn = (x.shape[0] - 1) * (x - x[0]) / (x[-1] - x[0])
-        weights = scipy.integrate.newton_cotes(rn)[0]
-        return (x[-1] - x[0]) / (x.shape[0] - 1) * np.dot(weights, f)
-
-    ret = 0
-    for idx in range(0, x.shape[0], p - 1):
-        ret += newton_cotes(x[idx : idx + p], f[idx : idx + p])
-    return ret
